requirements.txt
anyio==3.6.2
bcrypt==4.0.1
blinker==1.4
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
command-not-found==0.3
cryptography==39.0.0
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.3.0
email-validator==1.3.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.20.2
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.2
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
orjson==3.8.5
pycparser==2.21
pydantic==1.10.4
pygame==2.1.3
PyGObject==3.42.1
PyJWT==2.3.0
pyodbc==4.0.35
pyOpenSSL==23.0.0
pyparsing==2.4.7
python-apt==2.3.0
python-decouple==3.7
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==5.4.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
systemd-python==234
typing_extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
uvicorn==0.20.0
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.18.1
websockets==10.4
zipp==1.0.0
